chore(): Keep HTTP security configuration aside in order to extend it
