+++ packaging/local-dev/src/main/java/com/chutneytesting/SecSecurityMemoryConfig.java
@@ -1,6 +1,6 @@
 package com.chutneytesting;
 
-import com.chutneytesting.security.ChutneySecurityConfig;
+import com.chutneytesting.security.ChutneyHttpSecurityConfig;
 import com.chutneytesting.security.api.UserDto;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
@@ -10,7 +10,7 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 @Configuration
 @Profile("dev-auth")
 @Order(1)
-public class SecSecurityMemoryConfig extends ChutneySecurityConfig {
+public class SecSecurityMemoryConfig extends ChutneyHttpSecurityConfig {
 
     @Override
     protected void configure(final HttpSecurity http) throws Exception {
